Add comments, more info saved in outputs and changed plotting of outputs.
Details:
Add additional comments to jpopmodel.functions.R

Chang jpopmodel.R so that cc, birth rate and mortality for each jcu are also saved in the outputs (in the all.outputs dataframe).

Change analyse.results.R so that it plots the cc of each jcu on the population trajectory plots.